pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.0",
    "fastapi-users[oauth,sqlalchemy]>=15.0.1",
    "httpx-oauth>=0.16.1",
    "imagekitio>=4.2.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "supabase>=2.24.0",
    "uvicorn[standard]>=0.38.0",
]
